Rename midi package files
